dom/media/tests/mochitest/test_peerConnection_replaceTrack.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="mediaStreamPlayback.js"></script>
  <script type="application/javascript" src="pc.js"></script>
  <script type="application/javascript" src="templates.js"></script>
  <script type="application/javascript" src="turnConfig.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript;version=1.8">
  createHTML({
    bug: "1032839",
    title: "Replace video track",
    visible: true
  });

  function isSenderOfTrack(sender) {
    return sender.track == this;
  }

  // Test basically just verifies that success callback is called at this point

  var test;
  runNetworkTest(function () {
    test = new PeerConnectionTest();
    test.setMediaConstraints([{video: true}], [{video: true}]);
    test.chain.removeAfter("PC_REMOTE_CHECK_MEDIA_FLOW_PRESENT");
    var flowtest = test.chain.remove("PC_REMOTE_CHECK_MEDIA_FLOW_PRESENT");
    test.chain.append(flowtest);
    test.chain.append([["PC_LOCAL_REPLACE_VIDEOTRACK",
      function (test) {
        var stream = test.pcLocal._pc.getLocalStreams()[0];
        var track = stream.getVideoTracks()[0];
        var sender = test.pcLocal._pc.getSenders().find(isSenderOfTrack, track);
        ok(sender, "track has a sender");
        navigator.mozGetUserMedia({video:true, fake: true}, function(newStream) {
          var newtrack = newStream.getVideoTracks()[0];
          sender.replaceTrack(newtrack,
            function() {
              ok(true, "replaceTrack success callback is called");
              is(sender.track, newtrack, "sender.track has been replaced");
              test.next();
            },
            function(err) {
              ok(false, "replaceTrack failed with error = " + err);
              test.next();
            });
        },
        function(err) {
          ok(false, "mozGetUserMedia failed. error = " + err);
          test.next();
        });
      }
    ]]);
    test.chain.append(flowtest);

    test.run();
  });
</script>
</pre>
</body>
</html>
